backend/requirements.txt
asgiref==3.8.1
dj-rest-auth==2.1.9
Django==5.1.2
django-allauth==0.50.0
django-cors-headers==4.4.0
django-filter==24.3
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
oauthlib==3.2.2
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2024.2
requests-oauthlib==2.0.0
sqlparse==0.5.1
